Added Watch Item
When an item is listed, either through running a search or by live
search, there is now a 'watch' button that will allow you to keep track
of items even after refreshing the page.  These can be accessed via a
new tab at the bottom of the page.  Watched items can be removed by
clicking the 'Stop Watching' button on the Watched tab.